Move WHOIS code to libclients.la
