--- src/tests/stub_whois.cc
@@ -1,17 +1,0 @@
-/*
- * Copyright (C) 1996-2023 The Squid Software Foundation and contributors
- *
- * Squid software is distributed under GPLv2+ license and includes
- * contributions from numerous individuals and organizations.
- * Please see the COPYING and CONTRIBUTORS files for details.
- */
-
-#include "squid.h"
-#include "FwdState.h"
-#include "whois.h"
-
-#define STUB_API "whois.cc"
-#include "tests/STUB.h"
-
-void whoisStart(FwdState *) STUB
-
